requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
py-cord @ git+https://github.com/Pycord-Development/pycord
python-dotenv==0.21.0
yapf==0.32.0
yarl==1.8.1
json5==0.9.10
dataclass-wizard==0.22.2
aiofiles==22.1.0
thefuzz[speedup]==0.19.0
